Initial commit
This serves a simple-as-possible version locally. To run, install
dependencies (flask, flask-wtf), cd into top level directory (where
README.markdown is), and run

    FLASK_ENV=development FLASK_APP=app/main.py flask run

The "development" env indicates that flask should run in debug mode. If
desired, this can be served over the local network by appending

    --host=0.0.0.0

to the end of the previous command.

Next steps involve improving the is-this-a-sport function and making
anything look not terrible.